Update statistics to note the stats methodology change
This patch updates the statistics pages to point to the new mechanism for
generating statistics (ct-mapreduce), link to the Community post, and
annotates the cut-over date on the statistics charts.

The annotation I think we'd want to eventually remove, or maybe make less
obvious / hide in some way, but for now I think blunt is appropriate.

Relatedly, this patch updates the "backup" cert-timeline.tsv to current-date
and catches the methodology change.

This patch can merge at any time, but please coordinate with me before
updating the website so I can also cut-over the generation from the old
way to the new way.